rename home components + delete unused stuff
