+++ src/lib/Containers/index.tsx
@@ -5,7 +5,7 @@ import CitySavedList from "lib/Scenes/City/CitySavedList"
 import CitySectionList from "lib/Scenes/City/CitySectionList"
 import { PartnerContainer as Partner } from "lib/Scenes/Partner"
 import Fair from "../Scenes/Fair"
-import Home from "../Scenes/Home"
+import { Home } from "../Scenes/Home/Home"
 import MyProfile from "../Scenes/Settings/MyProfile"
 import Show from "../Scenes/Show"
 import Artist from "./Artist"
